refactor(ads131m08): move typed config fields into config::fields
